chore(136): split tests/test_conventions.py into 5 files by feature
Split the 2,847-line test_conventions.py into:
- test_conventions_layers.py (366 lines) — layer detection, naming pattern inference
- test_conventions_inference.py (676 lines) — reference rules, common columns, column style, run_inference, get_conventions
- test_conventions_overrides.py (442 lines) — YAML overrides, config discovery, CLI (init/refresh/diff)
- test_conventions_tagging.py (812 lines) — clustering, tag confidence, tag storage, tag queries
- test_conventions_placement.py (593 lines) — find_similar_models, suggest_placement

All 91 tests still pass. No behavior changes.
